refactor(host-broker): extract broker tests
